moduleFactory: Support slot validations
Until now,  we could put whatever module  in whatever slot,
and  let the app break in  "mysterious ways".  The idea is
to allow modules to declare what interfaces or classes are
required for each slot, enforce those requirements and warn
the app developer with meaningful message.

A slot definition looks like this:

    Slots: {
        'card': {
            requires: [
              Card.Card,
            ],
        },
    },

A list of required interfaces/classes can be provided
for each slott. Those will be enforced by the module
factory.

A special keyword 'self' is added to  cover cases where
the  requirement  is the same class  where the slot was
defined. e.g. a Filter interface defining a 'sub-filter'
slot:

    Slots: {
        'sub-filter': {
            requires: [
              'self',
            ],
        },
    },

https://phabricator.endlessm.com/T22786