modules: Properly display global-search items in course apps
Course apps display video and audio articles using the
lightbox. The current CourseHistoryStore implementation
did not take this behavior into account, when loading
items from the DBUS interface.

Instead, it used the default behavior of opening all
articles on the article page. So, write a proper DBUS
interface handler for the CourseHistoryStore.

https://phabricator.endlessm.com/T23305