application: Expose DBus interface to restart the app
This exposes a com.endlessm.KnowledgeControl interface with a
Restart(paths: a{sh}, resources: ah, shards: ah, options: a{sv}) method
that can be used to restart the app with different command line arguments.

Currently the arguments with paths (-O, -w, -J) are supported,
as keys in the "paths" dictionary ('theme', 'webcss', 'ui' respectively.)
The values of these keys must be handles.

The extra GResource argument (-E) is supported as well, although since
it can be specified multiple times, it is passed as an array of handles,
as a separate argument.

The content-path argument (-p) is a bit more difficult since it requires
a whole directory. To specify a content path, pass a handle to the
manifest.json file as the 'manifest' key in the "paths" dictionary, and
pass handles to the shard files in the "shards" argument. It is illegal
to pass a manifest without passing shards, or vice versa.

Other arguments can be supported in the future, either as additional
entries in the "paths" dictionary, or using keys in the "options"
dictionary (which is currently ignored as no options are currently
supported.)

This also contains several workarounds: one for a flatpak-spawn bug which
spawned all apps with GIO_USE_VFS=local in the environment, which blocked
our EknVfs from being loaded; a workaround for a GJS bug which made strings
show up as zero-length on the other side of DBus
(https://gitlab.gnome.org/GNOME/gjs/issues/203) and a workaround for
GJS's lack of the ability to call DBus methods with GUnixFDLists
(https://gitlab.gnome.org/GNOME/gjs/issues/204).

https://phabricator.endlessm.com/T23773