Allow to show content in mediaLightbox that is not children of original model
The idea is to allow to load linked content like Audio and Video without
the necessity of having those as a children ("resources") of the
original article model.

That means that by default, if the article doesn't have children, the
linked model will add itself as item context so the lightbox can come up
with only one asset.