Remove custom autocomplete popover style
The search autocomplete popover appeared broken with the removal of the
style reset in EosWindow. This commit changes the search autocomplete
popover to also reset existing styles before applying its own style.
Unfortunately it is not possible to change the popover style directly
because this is specified in CSS bundled inside each knowledge app.

https://phabricator.endlessm.com/T21463